# docs/openapi.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "EkMaap onion grading API",
  "description": "SIH26031 - photo-based onion grading. Prototype: see docs/ for limits.",
  "version": "0.2.0"
 },
 "paths": {
...
  "/api/v1/auth/login": {
   "post": {
    "tags": [
     "auth"
    ],
    "summary": "Login",
    "operationId": "login_api_v1_auth_login_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/Login"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/auth/me": {
   "get": {
    "tags": [
     "auth"
    ],
    "summary": "Me",
    "operationId": "me_api_v1_auth_me_get",
    "parameters": [
     {
      "name": "authorization",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Authorization"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/centres": {
   "post": {
    "tags": [
     "admin"
    ],
    "summary": "Create Centre",
    "operationId": "create_centre_api_v1_centres_post",
    "parameters": [
     {
      "name": "authorization",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Authorization"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/CentreIn"
       }
      }
     }
    },
    "responses": {
     "201": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   },
   "get": {
    "tags": [
     "admin"
    ],
    "summary": "List Centres",
    "operationId": "list_centres_api_v1_centres_get",
    "parameters": [
     {
      "name": "authorization",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Authorization"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/users": {
   "post": {
    "tags": [
     "admin"
    ],
    "summary": "Create User",
    "operationId": "create_user_api_v1_users_post",
    "parameters": [
     {
      "name": "authorization",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Authorization"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/UserIn"
       }
      }
     }
    },
    "responses": {
     "201": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   },
   "get": {
    "tags": [
     "admin"
    ],
    "summary": "List Users",
    "operationId": "list_users_api_v1_users_get",
    "parameters": [
     {
      "name": "authorization",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Authorization"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/farmers": {
   "post": {
    "tags": [
     "admin"
    ],
    "summary": "Create Farmer",
    "operationId": "create_farmer_api_v1_farmers_post",
    "parameters": [
     {
      "name": "authorization",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Authorization"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/FarmerIn"
       }
      }
     }
    },
    "responses": {
     "201": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   },
   "get": {
    "tags": [
     "admin"
    ],
    "summary": "List Farmers",
    "operationId": "list_farmers_api_v1_farmers_get",
    "parameters": [
     {
      "name": "q",
      "in": "query",
      "required": false,
      "schema": {
       "type": "string",
       "default": "",
       "title": "Q"
      }
     },
     {
      "name": "authorization",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Authorization"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/rulesets": {
   "get": {
    "tags": [
     "rule sets"
    ],
    "summary": "List Rule Sets",
    "operationId": "list_rule_sets_api_v1_rulesets_get",
    "parameters": [
     {
      "name": "authorization",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Authorization"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   },
   "post": {
    "tags": [
     "rule sets"
    ],
    "summary": "Create Rule Set",
    "operationId": "create_rule_set_api_v1_rulesets_post",
    "parameters": [
     {
      "name": "authorization",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Authorization"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/RuleSetIn"
       }
      }
     }
    },
    "responses": {
     "201": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/rulesets/active": {
   "get": {
    "tags": [
     "rule sets"
    ],
    "summary": "Get Active Rule Set",
    "operationId": "get_active_rule_set_api_v1_rulesets_active_get",
    "parameters": [
     {
      "name": "authorization",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Authorization"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/rulesets/{rule_set_id}/activate": {
   "post": {
    "tags": [
     "rule sets"
    ],
    "summary": "Activate Rule Set",
    "operationId": "activate_rule_set_api_v1_rulesets__rule_set_id__activate_post",
    "parameters": [
     {
      "name": "rule_set_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "format": "uuid",
       "title": "Rule Set Id"
      }
     },
     {
      "name": "authorization",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Authorization"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
 },
 "components": {
  "schemas": {
...
   "CentreIn": {
    "properties": {
     "code": {
      "type": "string",
      "maxLength": 32,
      "minLength": 1,
      "title": "Code"
     },
     "name": {
      "type": "string",
      "title": "Name"
     },
     "district": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "District"
     },
     "state": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "State"
     },
     "lat": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Lat"
     },
     "lon": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Lon"
     }
    },
    "type": "object",
    "required": [
     "code",
     "name"
    ],
    "title": "CentreIn"
   },
...
   "FarmerIn": {
    "properties": {
     "name": {
      "type": "string",
      "title": "Name"
     },
     "phone": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Phone"
     },
     "village": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Village"
     },
     "district": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "District"
     },
     "consent": {
      "type": "boolean",
      "title": "Consent",
      "description": "Farmer agreed to their name/phone being stored"
     }
    },
    "type": "object",
    "required": [
     "name",
     "consent"
    ],
    "title": "FarmerIn"
   },
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "type": "array",
      "title": "Detail"
     }
    },
    "type": "object",
    "title": "HTTPValidationError"
   },
...
   "Login": {
    "properties": {
     "login": {
      "type": "string",
      "title": "Login"
     },
     "password": {
      "type": "string",
      "title": "Password"
     }
    },
    "type": "object",
    "required": [
     "login",
     "password"
    ],
    "title": "Login"
   },
...
   "RuleSet": {
    "properties": {
     "name": {
      "type": "string",
      "title": "Name",
      "default": "Demo rule set - 2026 news values, NOT official"
     },
     "version": {
      "type": "string",
      "title": "Version",
      "default": "demo-2026-09-28"
     },
     "grade_a_min_mm": {
      "type": "number",
      "exclusiveMinimum": 0.0,
      "title": "Grade A Min Mm",
      "default": 45
     },
     "urs_min_mm": {
      "type": "number",
      "exclusiveMinimum": 0.0,
      "title": "Urs Min Mm",
      "default": 35
     },
     "grade_a_max_mm": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Grade A Max Mm"
     },
     "map_sprouted": {
      "type": "string",
      "enum": [
       "Grade A",
       "URS",
       "Reject"
      ],
      "title": "Map Sprouted",
      "default": "Reject"
     },
     "map_damaged": {
      "type": "string",
      "enum": [
       "Grade A",
       "URS",
       "Reject"
      ],
      "title": "Map Damaged",
      "default": "URS"
     },
     "map_rotten": {
      "type": "string",
      "enum": [
       "Grade A",
       "URS",
       "Reject"
      ],
      "title": "Map Rotten",
      "default": "Reject"
     },
     "review_below_confidence": {
      "type": "number",
      "maximum": 1.0,
      "minimum": 0.0,
      "title": "Review Below Confidence",
      "default": 0.6
     },
     "basis": {
      "type": "string",
      "const": "count",
      "title": "Basis",
      "default": "count"
     }
    },
    "type": "object",
    "title": "RuleSet"
   },
   "RuleSetIn": {
    "properties": {
     "params": {
      "$ref": "#/components/schemas/RuleSet"
     },
     "source_note": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Source Note"
     },
     "activate": {
      "type": "boolean",
      "title": "Activate",
      "default": false
     }
    },
    "type": "object",
    "required": [
     "params"
    ],
    "title": "RuleSetIn"
   },
...
   "UserIn": {
    "properties": {
     "name": {
      "type": "string",
      "title": "Name"
     },
     "login": {
      "type": "string",
      "title": "Login"
     },
     "password": {
      "type": "string",
      "minLength": 8,
      "title": "Password"
     },
     "role": {
      "type": "string",
      "pattern": "^(officer|supervisor|labeller|admin)$",
      "title": "Role"
     },
     "centre_id": {
      "anyOf": [
       {
        "type": "string",
        "format": "uuid"
       },
       {
        "type": "null"
       }
      ],
      "title": "Centre Id"
     }
    },
    "type": "object",
    "required": [
     "name",
     "login",
     "password",
     "role"
    ],
    "title": "UserIn"
   },
   "ValidationError": {
    "properties": {
     "loc": {
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ]
      },
      "type": "array",
      "title": "Location"
     },
     "msg": {
      "type": "string",
      "title": "Message"
     },
     "type": {
      "type": "string",
      "title": "Error Type"
     },
     "input": {
      "title": "Input"
     },
     "ctx": {
      "type": "object",
      "title": "Context"
     }
    },
    "type": "object",
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "title": "ValidationError"
   }
  }
 }
}